UI: Add LDCListWidget - QListWidget subclass
The behaviour of QListWidget is to process by default
any mouse buttons double clicks. This can cause, in certain
situations/usage cases, undesirable results.

This subclass filters out all mouse buttons double clicks but
the left one.

As an example: when doing a right mouse button double click on
an 'sources list' item, it is possible to have both the
properties window, and the context menu shown.
Not pretty at all.